backend_django/requirements.txt
django
pymysql
python-dotenv
bcrypt
